Dopracowanie frontendu. Wyłączenie autoryzacji i test wydajności przy 200 użytkownikach

+++ Backend/AssignmentsSystem/src/main/java/pl/edu/pwr/AssignmentsSystem/Powierzenia/Entrypoints/PowierzenieController.java
@@ -4,11 +4,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.edu.pwr.AssignmentsSystem.Commons.Auth.AuthenticationService;
+import pl.edu.pwr.AssignmentsSystem.Commons.Auth.ROLE;
 import pl.edu.pwr.AssignmentsSystem.Commons.Dto.IdentyfikatorSemestruDto;
 import pl.edu.pwr.AssignmentsSystem.Commons.Dto.PlanPowierzenDto;
 import pl.edu.pwr.AssignmentsSystem.Commons.Dto.PowierzenieDto;
 import pl.edu.pwr.AssignmentsSystem.Powierzenia.Usecase.PowierzenieService;
 
+import java.util.ArrayList;
 import java.util.List;
 @CrossOrigin
 @RestController
@@ -17,15 +20,24 @@ public class PowierzenieController {
     @Autowired
     private PowierzenieService powierzenieService;
 
+    @Autowired
+    private AuthenticationService authenticationService;
+
     @PostMapping("/getPlanPowierzen")
     public List<PlanPowierzenDto> getPlanPowierzen(@RequestBody IdentyfikatorSemestruDto identyfikatorSemestruDto)
     {
-        return powierzenieService.getAllPlanPowerzen(identyfikatorSemestruDto);
+        if(authenticationService.isUserPermitted(ROLE.USER.name())) {
+            return powierzenieService.getAllPlanPowerzen(identyfikatorSemestruDto);
+        }
+        return new ArrayList<>();
     }
 
     @PutMapping("/savePowierzenie")
     public ResponseEntity savePowierzenie(@RequestBody PlanPowierzenDto planPowierzenDto)
     {
+        if(!authenticationService.isUserPermitted(ROLE.ADMIN.name())) {
+            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+        }
         try {
             if(powierzenieService.savePowierzenie(planPowierzenDto))
             {
@@ -39,18 +51,27 @@ public class PowierzenieController {
     @PutMapping("/savePlanPowierzen")
     public ResponseEntity savePlanPowierzen(@RequestBody PlanPowierzenDto planPowierzenDto)
     {
+        if(!authenticationService.isUserPermitted(ROLE.ADMIN.name())) {
+            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+        }
         powierzenieService.savePlanPowierzen(planPowierzenDto);
         return new ResponseEntity(HttpStatus.CREATED);
     }
     @PostMapping("/removePowierzenie")
     public ResponseEntity removePowierzenie(@RequestBody PowierzenieDto powierzenieDto)
     {
+        if(!authenticationService.isUserPermitted(ROLE.ADMIN.name())) {
+            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+        }
         powierzenieService.removePowierzenie(powierzenieDto);
         return new ResponseEntity(HttpStatus.OK);
     }
     @PostMapping("/approvePlanPowierzen")
     public ResponseEntity approvePlanPowierzen(@RequestBody PlanPowierzenDto planPowierzenDto)
     {
+        if(!authenticationService.isUserPermitted(ROLE.ADMIN.name())) {
+            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+        }
         List<String> approvePlanPowierzen = powierzenieService.approvePlanPowierzen(planPowierzenDto);
         if(approvePlanPowierzen.isEmpty()) {
             return new ResponseEntity(HttpStatus.OK);
@@ -60,6 +81,9 @@ public class PowierzenieController {
     @PostMapping("/returnPlanPowierzen")
     public ResponseEntity returnPlanPowierzen(@RequestBody PlanPowierzenDto planPowierzenDto)
     {
+        if(!authenticationService.isUserPermitted(ROLE.ADMIN.name())) {
+            return new ResponseEntity(HttpStatus.UNAUTHORIZED);
+        }
         if(powierzenieService.returnPlanPowierzen(planPowierzenDto)){
             return new ResponseEntity(HttpStatus.OK);
         }
